+++ Source/FirstParty/Src/gui/InputBox.hpp
@@ -46,7 +46,7 @@ public:
 private:
     std::unique_ptr<MenuElement> doClone() const override;
     void doDraw(const DrawParameter& params) override;
-    void updated(const sf::RenderWindow& screen, const float time, const sf::Vector2i& mouseOffset = sf::Vector2i(0, 0)) override;
+    void updated(const sf::RenderWindow& screen, const double time, const sf::Vector2i& mouseOffset = sf::Vector2i(0, 0)) override;
     void layoutUpdated(const sf::Vector2f& screenSize) override;
 
     sf::RectangleShape m_backgroundShade;
