+++ Source/FirstParty/Src/gui/InteractiveLabel.hpp
@@ -15,7 +15,7 @@ private:
 
     std::unique_ptr<MenuElement> doClone() const override;
     void onDrawAdditionalForeground(const DrawParameter& params) override;
-    void updated(const sf::RenderWindow& screen, const float time, const sf::Vector2i& mouseOffset = sf::Vector2i(0, 0)) override;
+    void updated(const sf::RenderWindow& screen, const double time, const sf::Vector2i& mouseOffset = sf::Vector2i(0, 0)) override;
 
 public:
     InteractiveLabel(const std::string& text,
