+++ Source/FirstParty/Src/gui/KeyboardMenu.hpp
@@ -13,7 +13,7 @@ private:
     int m_lastKey;
 
     void doDraw(const DrawParameter& params) override;
-    void updated(sf::RenderWindow& screen, const float time) override;
+    void updated(sf::RenderWindow& screen, const double time) override;
 
 public:
     static const int BUTTON_KEYBOARD_OFF = 99;
